+++ js/src/analysis/PostDataCollector.js
@@ -10,7 +10,7 @@ import { update as updateTrafficLight } from "../ui/trafficLight";
 import { update as updateAdminBar }from "../ui/adminBar";
 
 import publishBox from "../ui/publishBox";
-import isGutenbergDataAvailable from "../helpers/isGutenbergDataAvailable";
+import { isGutenbergDataAvailable } from "../helpers/isGutenbergAvailable";
 
 let $ = jQuery;
 let currentKeyword = "";
