+++ js/src/helpers/isGutenbergAvailable.js
@@ -7,11 +7,9 @@ import isUndefined from "lodash/isUndefined";
  *
  * @returns {boolean} True if the data API is available.
  */
-const isGutenbergDataAvailable = () => {
+export const isGutenbergDataAvailable = () => {
 	return ( ! isUndefined( window.wp ) && ! isUndefined( wp.data ) );
 };
-
-export default isGutenbergDataAvailable;
 
 /**
  * Checks if the Gutenberg editor is being used.
